+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.Storage;
+import duke.task.TaskList;
+
 public abstract class Command {
     private String command;
