+++ src/main/java/duke/command/deleteCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.Ui;
+
 public class deleteCommand extends Command {
     private String command;
