+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,14 @@
+package duke.exception;
+
+public class DukeException extends Exception {
+    private String message;
+    
+    public DukeException(String message) {
+        this.message = message;
+    }
+    
+    @Override
+    public String toString() {
+        return message;
+    }
+}
