+++ src/main/java/duke/exception/EmptyDescriptionException.java
@@ -1,0 +1,12 @@
+package duke.exception;
+
+public class EmptyDescriptionException extends DukeException {
+    public EmptyDescriptionException(String message) {
+        super(message);
+    }
+    
+    @Override
+    public String toString() {
+        return super.toString();
+    }
+}
