+++ src/main/java/duke/exception/InvalidDateTimeException.java
@@ -1,0 +1,11 @@
+package duke.exception;
+
+public class InvalidDateTimeException extends DukeException {
+    public InvalidDateTimeException(String message) {
+        super(message);
+    }
+    
+    public String toString() {
+        return super.toString();
+    }
+}
