+++ src/main/java/duke/task/TaskList.java
@@ -1,3 +1,6 @@
+package duke.task;
+
+import duke.Ui;
 import java.util.ArrayList;
 
 public class TaskList {
